+++ changelog.md
@@ -1,6 +1,9 @@
 ## 1.0.0
 - Full Release
 
+## 1.0.1
+- Fixed issue with ReadOnlyList returning undefined for items when using `match()`
+
 ## 1.1.0
 - Enhancers can now modify an entity and changes will be tracked
 - Added transactions to bulkOperations for DB Plugins.  Now devs can see what properties have changed